Fix in DBHandler, Anpassung Recycler + ArrayListAdapter

+++ app/src/main/java/com/example/mobappproject/recycleViewIngredients/RecyclerAdapter.kt
@@ -3,11 +3,17 @@ package com.example.mobappproject.recycleViewIngredients
 import android.content.Context
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import android.widget.Button
 import androidx.recyclerview.widget.RecyclerView
 import com.example.mobappproject.database.DBIngredient
 import com.example.mobappproject.database.DatabaseHandler
 
-class RecyclerAdapter(private val list: ArrayList<DBIngredient>, private val adapter: ArrayListAdapter, private val db: DatabaseHandler) : RecyclerView.Adapter<IngredientHolder>() {
+/**
+ * @param list Contains ArrayList that should be displayed
+ * @param adapter ArrayListAdapter that contains the available ingredients for the AutoCompleteTextView
+ * @param db DatabseHanlder to remove the ingredients from the Userlist. Does noting if null
+ */
+class RecyclerAdapter(private val list: ArrayList<DBIngredient>, private val adapter: ArrayListAdapter, private val db: DatabaseHandler?) : RecyclerView.Adapter<IngredientHolder>() {
 
     private var ingredient: DBIngredient? = null
 
@@ -28,7 +34,9 @@ class RecyclerAdapter(private val list: ArrayList<DBIngredient>, private val ada
 
     fun remove(position: Int) {
         val ing = list[position]
-        db.removeStoreIngredient(ing)
+        if(db != null) {
+            db.removeStoreIngredient(ing)
+        }
         list.removeAt(position)
         if(position != 0)
             notifyItemRangeChanged(position, list.size)
